# portal/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Runner Portal API",
    "description": "API do portal de assessorias esportivas Omni Runner. Gerencia atletas, treinos, financeiro, OmniCoins e administração da plataforma.",
    "version": "1.0.0",
    "contact": {
      "name": "Omni Runner"
    }
  },
  "servers": [
    {
      "url": "https://portal.omnirunner.app",
      "description": "Produção"
    }
  ],
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/api/gateway-preference": {
      "get": {
        "tags": ["Settings"],
        "summary": "Obter gateway preferido",
        "description": "Retorna o gateway de pagamento preferido (mercadopago ou stripe).",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "preferred_gateway": { "type": "string", "enum": ["mercadopago", "stripe"] } }
                }
              }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Definir gateway preferido",
        "description": "Define o gateway de pagamento preferido para a assessoria.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["preferred_gateway"],
                "properties": {
                  "preferred_gateway": { "type": "string", "enum": ["mercadopago", "stripe"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "preferred_gateway": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin_master" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/export/financial": {
      "get": {
        "tags": ["Export", "Financial"],
        "summary": "Exportar financeiro CSV",
        "description": "Exporta ledger financeiro da assessoria em CSV.",
        "responses": {
          "200": {
            "content": {
              "text/csv": { "schema": { "type": "string" } }
            }
          },
          "400": { "description": "Grupo não selecionado" },
          "401": { "description": "Não autorizado" },
          "403": { "description": "Apenas admin/coach" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/fees": {
      "get": {
        "tags": ["Platform Admin"],
        "summary": "Listar taxas",
        "description": "Lista configuração de taxas da plataforma.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "fees": { "type": "array", "items": { "type": "object" } } }
                }
              }
            }
          },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" }
        }
      },
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Atualizar taxa",
        "description": "Atualiza taxa (clearing, swap, maintenance, billing_split).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fee_type"],
                "properties": {
                  "fee_type": { "type": "string", "enum": ["clearing", "swap", "maintenance", "billing_split"] },
                  "rate_pct": { "type": "number", "minimum": 0, "maximum": 100 },
                  "rate_usd": { "type": "number", "minimum": 0, "maximum": 10 },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Taxa atualizada" },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/assessorias": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Aprovar/rejeitar/suspender assessoria",
        "description": "Altera status de aprovação de uma assessoria.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "group_id"],
                "properties": {
                  "action": { "type": "string", "enum": ["approve", "reject", "suspend"] },
                  "group_id": { "type": "string", "format": "uuid" },
                  "reason": { "type": "string", "maxLength": 500 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["approved", "rejected", "suspended"] },
                    "group_id": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
    "/api/platform/feature-flags": {
      "post": {
        "tags": ["Platform Admin"],
        "summary": "Atualizar feature flag",
        "description": "Atualiza enabled e rollout_pct de uma feature flag.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "enabled", "rollout_pct"],
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "enabled": { "type": "boolean" },
                  "rollout_pct": { "type": "integer", "minimum": 0, "maximum": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Feature flag atualizada" },
          "400": { "description": "Validação falhou" },
          "401": { "description": "Não autenticado" },
          "403": { "description": "Não é admin da plataforma" },
          "429": { "description": "Rate limit excedido" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT do Supabase (access_token da sessão)"
      }
    },
...
  }
}